Refactored movies table - created tableHeader component for reusable sorting
